requirements.txt
absl-py==0.11.0
astor==0.8.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.4.1
Bottleneck==1.3.2
brotlipy==0.7.0
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1606255098796/work
chardet @ file:///tmp/build/80754af9/chardet_1605303175790/work
click==7.1.2
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018143336/work
cryptography @ file:///tmp/build/80754af9/cryptography_1607635303852/work
cycler==0.10.0
cymem @ file:///tmp/build/80754af9/cymem_1605062275754/work
dataclasses==0.7
decorator==4.4.2
dnspython==2.0.0
fastai==1.0.61
fastcore @ file:///home/jhoward/anaconda3/conda-bld/fastcore_1607784119732/work
fastprogress @ file:///home/jhoward/anaconda3/conda-bld/fastprogress_1597927594295/work
Flask==1.1.2
Flask-Cors==3.0.10
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1602276842396/work
ipykernel==5.4.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib @ file:///tmp/build/80754af9/joblib_1607970656719/work
jsonschema==3.0.2
jupyter-client==6.1.7
jupyter-core==4.7.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014557000/work
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603373816490/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mongoengine==0.22.1
murmurhash @ file:///tmp/build/80754af9/murmurhash_1607456103988/work
numexpr==2.7.1
numpy==1.16.4
nvidia-ml-py3==7.352.0
olefile==0.46
packaging @ file:///tmp/build/80754af9/packaging_1607971725249/work
pandas==1.1.5
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1603822238230/work
plac @ file:///tmp/build/80754af9/plac_1594261902054/work
preshed==3.0.2
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.7.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
pyvi==0.1
PyYAML==5.3.1
pyzmq==20.0.0
requests @ file:///tmp/build/80754af9/requests_1606691187061/work
ruamel-yaml==0.15.87
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1598379690588/work
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
six @ file:///tmp/build/80754af9/six_1605205335545/work
sklearn-crfsuite==0.3.6
soupsieve==2.1
spacy @ file:///tmp/build/80754af9/spacy_1600554229564/work
srsly @ file:///tmp/build/80754af9/srsly_1607548540192/work
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thinc @ file:///tmp/build/80754af9/thinc_1594251955397/work
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm @ file:///tmp/build/80754af9/tqdm_1607369919789/work
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
wasabi @ file:///tmp/build/80754af9/wasabi_1599244993840/work
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1604001098328/work
